updates server example to load paramters from args

+++ src/irsol/include/irsol/server/server.hpp
@@ -2,20 +2,21 @@
 
 #include <poll.h>
 #include <vector>
+#include <stdint.h>
 
 namespace irsol {
 
 class Server {
 public:
-    Server(int port, int maxClients);
+    Server(uint16_t port, uint16_t maxClients);
     ~Server();
 
     void run();
     void terminate();
 
 private:
-    int m_port;
-    int m_maxClients;
+    uint16_t m_port;
+    uint16_t m_maxClients;
     int m_serverSocket;
     bool m_terminate;
     std::vector<struct pollfd> m_fds;
